feat(room): 用可回收 uint 生成器统一 roomId
业务层 roomId 改为 uint；协议 int64 仅在边界转换；关房后可回收复用。

+++ Entity/Runtime/room/RoomFrameSync.cs
@@ -11,10 +11,10 @@ namespace Entity.Runtime.room;
 /// </summary>
 public sealed class RoomFrameSync
 {
-    private readonly Func<long> _getRoomId;
+    private readonly Func<uint> _getRoomId;
     private readonly RoomFrameWindow _frameWindow;
 
-    public RoomFrameSync(Func<long> getRoomId)
+    public RoomFrameSync(Func<uint> getRoomId)
     {
         _getRoomId = getRoomId ?? throw new ArgumentNullException(nameof(getRoomId));
         _frameWindow = new RoomFrameWindow(RoomConfig.FrameBufferCapacity);
