+++ Hotfix/Scene/Rooms/Handler/CreateAndEntry.cs
@@ -39,7 +39,7 @@ public sealed class RoomsCreateAndEntryHandler
 
     private static long TryGetRoomId(InnerResult result)
     {
-        if (result.Args is { Count: > 0 } && result.Args[0] is long roomId)
+        if (result.Args is { Count: > 0 } && result.Args[0] is uint roomId)
         {
             return roomId;
         }
